requirements.txt
aiohttp
aiosignal
annotated-types
anyio
asgiref
async-timeout
attrs
boto3
botocore
certifi
charset-normalizer
dataclasses-json
distro
Django
django-cors-headers
django-storages
djangorestframework
djangorestframework-simplejwt
exceptiongroup
faiss-cpu
filelock
frozenlist
fsspec
gunicorn
h11
httpcore
httpx
huggingface-hub
idna
install
Jinja2
jmespath
joblib
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-text-splitters
langsmith
MarkupSafe
marshmallow
mpmath
multidict
mypy-extensions
networkx
numpy
openai
orjson
packaging
pillow
pydantic
pydantic_core
PyJWT
PyMuPDF
PyMuPDFb
python-dateutil
python-dotenv
PyYAML
regex
requests
s3transfer
safetensors
scikit-learn
scipy
sentence-transformers
six
sniffio
SQLAlchemy
sqlparse
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
torch
tqdm
transformers
typing-inspect
typing_extensions
urllib3
yarl
